refactor: 💡 SQForm to ts
